RolesIdmResourceTest test is extracted
